+++ lib/slack-ruby-bot/app.rb
@@ -8,6 +8,7 @@ module SlackRubyBot
     def initialize
       SlackRubyBot.configure do |config|
         config.token = ENV['SLACK_API_TOKEN'] || fail("Missing ENV['SLACK_API_TOKEN'].")
+        config.aliases = ENV['SLACK_RUBY_BOT_ALIASES'].split(' ') if ENV['SLACK_RUBY_BOT_ALIASES']
       end
       Slack.configure do |config|
         config.token = SlackRubyBot.config.token