+++ lib/slack-ruby-bot/config.rb
@@ -4,10 +4,18 @@ module SlackRubyBot
 
     attr_accessor :token
     attr_accessor :url
+    attr_accessor :aliases
     attr_accessor :user
     attr_accessor :user_id
     attr_accessor :team
     attr_accessor :team_id
-    attr_accessor :secret
+
+    def names
+      [user, aliases].compact.flatten
+    end
+
+    def name?(name)
+      name && names.include?(name.downcase)
+    end
   end
 end